Data receiving for all journals

+++ app/src/main/java/com/example/diplomandroid/ui/band/BandViewModel.java
@@ -57,7 +57,7 @@ public class BandViewModel extends AndroidViewModel {
         repository.saveCalories(context, caloriesAmount, new Date().toString());
     }
 
-    public void saveWeight(Context context, Double weight) {
+    public void saveWeight(Context context, Integer weight) {
         repository.saveWeight(context, weight, new Date().toString());
     }
 }
